Adapter: refactor API models
