+++ adapter/src/adapter/api/DataImportResponse.dto.ts
