--- adapter/src/adapter/api/Endpoint.dto.ts
@@ -1,12 +1,0 @@
-export interface AdapterConfigDTO {
-  protocol: ProtocolConfigDTO;
-  format: FormatConfigDTO;
-}
-export interface ProtocolConfigDTO {
-  type: string;
-  parameters: Record<string, unknown>;
-}
-export interface FormatConfigDTO {
-  type: string;
-  parameters: Record<string, unknown>;
-}
