+++ adapter/src/adapter/exceptions/AdapterError.ts
