+++ adapter/src/adapter/exceptions/InterpreterParameterError.ts
