--- adapter/src/adapter/model/FormatConfig.ts
@@ -1,6 +1,0 @@
-import { Interpreter } from '../interpreter';
-
-export interface FormatConfig {
-  format: Interpreter;
-  parameters: Record<string, unknown>;
-}
